requirements.txt
# Medicaid FWA Analytics Pipeline — Python Dependencies

# Core data processing
duckdb>=0.9.0
numpy>=1.24.0
scipy>=1.10.0
pandas>=2.0.0

# Machine learning
scikit-learn>=1.3.0
xgboost>=2.0.0

# Deep learning
torch>=2.0.0

# Visualization
matplotlib>=3.7.0

# Configuration
PyYAML>=6.0

# Network analysis
networkx>=3.0

# Parallel execution
joblib>=1.3.0

# HTTP requests (LEIE, NPPES downloads)
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
